+++ apps/sess_id.c
@@ -114,7 +114,7 @@ int sess_id_main(int argc, char **argv)
             goto end;
         }
         if (!SSL_SESSION_set1_id_context(x, (unsigned char *)context,
-                                         ctx_len)) {
+                                         (unsigned int)ctx_len)) {
             BIO_printf(bio_err, "Error setting id context\n");
             goto end;
         }
